+++ src/OrderScrapper/OrderScrapper/Classes/OrdersExtractor.swift
@@ -54,6 +54,17 @@ public class OrdersExtractor {
         }
         //get scripts for the order sources
         BSScriptFileManager.shared.loadScriptFile()
+
+        _ = AmazonService.getConfigs() {configs, error in
+            if let configs = configs {
+                print("### Timeout ",configs.timeoutValue!)
+                if let timeoutValue = configs.timeoutValue {
+                    LibContext.shared.timeoutValue = timeoutValue
+                } else {
+                    LibContext.shared.timeoutValue = AppConstants.timeoutCounter
+                }
+            }
+        }
         
         isInitialized = true
     }